requirements.txt
numpy>=1.24
sounddevice>=0.4.6
openai>=1.31.0
python-dotenv>=1.0.1
TTS>=0.22.0
torch>=2.2.0
openai>=1.52.0,<2.0.0
numpy>=1.26.0,<3.0.0
sounddevice>=0.4.6,<0.5.0
python-dotenv>=1.0.0,<2.0.0
TTS>=0.22.0,<0.23.0
realtime-tts>=0.4.0,<0.5.0
